Intro to Redux Part 2
