+++ package.json
@@ -11,6 +11,7 @@
     "@fortawesome/fontawesome-svg-core": "^6.1.1",
     "@fortawesome/free-solid-svg-icons": "^6.1.1",
     "@fortawesome/react-fontawesome": "^0.1.18",
+    "@reduxjs/toolkit": "^1.8.2",
     "@types/react": "^18.0.9",
     "bluebird": "^3.7.2",
     "bootstrap": "5.0.0",
@@ -19,7 +20,7 @@
     "prop-types": "15.7.2",
     "react": "16.8.4",
     "react-dom": "16.8.4",
-    "react-redux": "6.0.1",
+    "react-redux": "7.1.0",
     "react-router-dom": "5.0.0",
     "react-toastify": "4.5.2",
     "redux": "4.0.1",
@@ -61,7 +62,7 @@
     "redux-immutable-state-invariant": "2.1.0",
     "redux-mock-store": "1.5.3",
     "rimraf": "2.6.3",
-    "style-loader": "0.23.1",
+    "style-loader": "^0.23.1",
     "webpack": "^5.72.1",
     "webpack-bundle-analyzer": "3.1.0",
     "webpack-cli": "^4.9.2",
