+++ src/css/bidpro.module.css
@@ -1,0 +1,19 @@
+.row {
+  display: flex !important;
+  align-items: center !important;
+  justify-content: center !important;
+}
+
+.value {
+  font-size: 32px !important;
+}
+
+.button {
+  width: 150px !important;
+  font-size: 32px !important;
+  padding: 1pm !important;
+  border: 2px solid rgb(0, 128, 128) !important;
+  color: rgb(0, 128, 128) !important;
+  background-color: white !important;
+  border-radius: 2px !important;
+}
