+++ src/redux/store.js
@@ -1,0 +1,8 @@
+import { configureStore } from "@reduxjs/toolkit";
+import counterReducer from "./counterSlice";
+
+export default configureStore({
+  reducer: {
+    counter: counterReducer,
+  },
+});
