+++ webpack.config.dev.js
@@ -37,7 +37,7 @@ module.exports = {
         use: ["babel-loader", "eslint-loader"],
       },
       {
-        test: /(\.css)$/,
+        test: /(\.css)$/i,
         use: ["style-loader", "css-loader"],
       },
       {
